requirements.txt
# Core dependencies for common development tasks
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Testing frameworks
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Utility libraries
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0

# Data processing
openpyxl>=3.1.0
pyyaml>=6.0

# Web scraping (optional)
beautifulsoup4>=4.12.0
selenium>=4.15.0

# API development (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
python-docx>=1.1.2
pywin32>=306